requirements.txt
pandas
numpy
matplotlib
plotly
scikit-learn
python-dotenv
pyarrow
streamlit